+++ ProducaoAPI/ProducaoAPI/Requests/ProcessoProducaoMateriaPrimaRequest.cs
@@ -1,3 +1,3 @@
 ﻿namespace ProducaoAPI.Requests;
 
-public record ProcessoProducaoMateriaPrimaRequest(int Id, double Quantidade);
+public record ProcessoProducaoMateriaPrimaRequest(int Id, decimal Quantidade);
